requirements.txt
absl-py==1.3.0
action-msgs==1.3.1
action-tutorials-interfaces==0.22.0
action-tutorials-py==0.22.0
actionlib-msgs==4.6.0
aio-pika==8.3.0
aiofiles==23.1.0
aiogram==2.23.1
aiohttp==3.8.4
aiormq==6.6.4
aiosignal==1.3.1
ament-cmake-test==1.5.1
ament-copyright==0.13.2
ament-cppcheck==0.13.2
ament-cpplint==0.13.2
ament-flake8==0.13.2
ament-index-python==1.5.1
ament-lint==0.13.2
ament-lint-cmake==0.13.2
ament-package==0.15.1
ament-pep257==0.13.2
ament-uncrustify==0.13.2
ament-xmllint==0.13.2
angles==1.16.0
APScheduler==3.9.1.post1
asgiref==3.6.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
bidict==0.22.1
boto3==1.26.86
botocore==1.29.86
builtin-interfaces==1.3.1
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorclass==2.2.2
coloredlogs==15.0.1
colorhash==1.2.1
composition-interfaces==1.3.1
confluent-kafka==1.9.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
cv-bridge==3.3.1
cycler==0.11.0
dask==2022.10.2
demo-nodes-py==0.22.0
diagnostic-msgs==4.6.0
Django==4.1.7
djangorestframework==3.14.0
dnspython==1.16.0
docopt==0.6.2
domain-coordinator==0.11.1
drf-yasg==1.21.5
example-interfaces==0.10.1
examples-rclpy-executors==0.16.1
examples-rclpy-minimal-action-client==0.16.1
examples-rclpy-minimal-action-server==0.16.1
examples-rclpy-minimal-client==0.16.1
examples-rclpy-minimal-publisher==0.16.1
examples-rclpy-minimal-service==0.16.1
examples-rclpy-minimal-subscriber==0.16.1
fbmessenger==6.0.0
fire==0.5.0
flatbuffers==23.3.3
fonttools==4.39.0
frozenlist==1.3.3
fsspec==2023.3.0
future==0.18.3
gast==0.5.3
geometry-msgs==4.6.0
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.51.3
h5py==3.8.0
httptools==0.5.0
humanfriendly==10.0
idna==3.4
image-geometry==3.3.1
inflection==0.5.1
interactive-markers==2.3.2
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpickle==2.2.0
jsonschema==4.16.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
laser-geometry==2.4.0
launch==1.2.0
launch-ros==0.21.0
launch-testing==1.2.0
launch-testing-ros==0.21.0
launch-xml==1.2.0
launch-yaml==1.2.0
libclang==15.0.6.1
lifecycle-msgs==1.3.1
locket==1.0.0
logging-demo==0.22.0
magic-filter==1.0.9
map-msgs==2.1.0
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.5.3
mattermostwrapper==2.2
message-filters==4.5.0
msgpack==1.0.4
multidict==5.2.0
nav-msgs==4.6.0
networkx==2.6.3
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
osrf-pycommon==2.1.0
packaging==20.9
pamqp==3.2.1
partd==1.3.0
pcl-msgs==1.0.0
pendulum-msgs==0.22.0
Pillow==9.4.0
pluggy==1.0.0
prompt-toolkit==3.0.28
protobuf==3.19.6
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
pydot==1.4.2
PyJWT==2.6.0
pykwalify==1.8.0
pymongo==3.10.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-engineio==4.3.4
python-qt-binding==1.2.0
python-socketio==5.7.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qt-dotgraph==2.3.1
qt-gui==2.3.1
qt-gui-cpp==2.3.1
qt-gui-py-common==2.3.1
quality-of-service-demo-py==0.22.0
questionary==1.10.0
randomname==0.1.5
rasa==3.4.4
rasa-sdk==3.4.1
rcl-interfaces==1.3.1
rclpy==3.7.0
rcutils==6.0.0
redis==4.5.1
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
resource-retriever==3.2.1
rmw-dds-common==1.7.1
rocketchat-API==1.28.1
ros2action==0.20.0
ros2bag==0.17.0
ros2cli==0.20.0
ros2component==0.20.0
ros2doctor==0.20.0
ros2interface==0.20.0
ros2launch==0.21.0
ros2lifecycle==0.20.0
ros2multicast==0.20.0
ros2node==0.20.0
ros2param==0.20.0
ros2pkg==0.20.0
ros2run==0.20.0
ros2service==0.20.0
ros2topic==0.20.0
rosbag2-interfaces==0.17.0
rosbag2-py==0.17.0
rosgraph-msgs==1.3.1
rosidl-adapter==3.3.0
rosidl-cli==3.3.0
rosidl-cmake==3.3.0
rosidl-generator-c==3.3.0
rosidl-generator-cpp==3.3.0
rosidl-generator-py==0.16.0
rosidl-parser==3.3.0
rosidl-pycommon==3.3.0
rosidl-runtime-py==0.11.0
rosidl-typesupport-c==2.2.0
rosidl-typesupport-cpp==2.2.0
rosidl-typesupport-fastrtps-c==2.4.0
rosidl-typesupport-fastrtps-cpp==2.4.0
rosidl-typesupport-introspection-c==3.3.0
rosidl-typesupport-introspection-cpp==3.3.0
rpyutils==0.3.1
rqt-action==2.1.0
rqt-bag==1.2.1
rqt-bag-plugins==1.2.1
rqt-console==2.1.0
rqt-graph==1.4.0
rqt-gui==1.2.0
rqt-gui-py==1.2.0
rqt-msg==1.3.0
rqt-plot==1.2.0
rqt-publisher==1.6.0
rqt-py-common==1.2.0
rqt-py-console==1.1.0
rqt-reconfigure==1.2.0
rqt-service-caller==1.1.0
rqt-shell==1.1.0
rqt-srv==1.1.0
rqt-topic==1.6.0
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
sanic==21.12.2
Sanic-Cors==2.0.1
sanic-jwt==1.8.0
sanic-routing==0.7.2
scikit-learn==1.1.3
scipy==1.8.1
sensor-msgs==4.6.0
sensor-msgs-py==4.6.0
sentry-sdk==1.11.1
shape-msgs==4.6.0
six==1.16.0
sklearn-crfsuite==0.3.6
slack-sdk==3.20.1
SQLAlchemy==1.4.46
sqlparse==0.4.3
sros2==0.11.1
statistics-msgs==1.3.1
std-msgs==4.6.0
std-srvs==4.6.0
stereo-msgs==4.6.0
tabulate==0.9.0
tarsafe==0.0.3
teleop-twist-keyboard==2.3.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.4
tensorflow-addons==0.17.1
tensorflow-estimator==2.8.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-text==2.8.2
termcolor==2.2.0
terminaltables==3.1.10
tf2-geometry-msgs==0.27.0
tf2-kdl==0.27.0
tf2-msgs==0.27.0
tf2-py==0.27.0
tf2-ros-py==0.27.0
tf2-sensor-msgs==0.27.0
tf2-tools==0.27.0
threadpoolctl==3.1.0
toolz==0.12.0
topic-monitor==0.22.0
tqdm==4.65.0
trajectory-msgs==4.6.0
turtlesim==1.5.0
twilio==7.14.2
typeguard==2.13.3
typing-utils==0.1.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
ujson==5.7.0
unique-identifier-msgs==2.3.1
uritemplate==4.1.1
urllib3==1.26.14
uvloop==0.17.0
visualization-msgs==4.6.0
wcwidth==0.2.6
webexteamssdk==1.6.1
websockets==10.4
Werkzeug==2.2.3
wrapt==1.15.0
yarl==1.8.2
